+++ x/concentrated-liquidity/client/cli/query.go
@@ -19,6 +19,9 @@ func GetQueryCmd() *cobra.Command {
 	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetClaimableIncentives)
 	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetIncentiveRecords)
 	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetCFMMPoolIdLinkFromConcentratedPoolId)
+	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetTickLiquidityNetInDirection)
+	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetPoolAccumulatorRewards)
+	osmocli.AddQueryCmd(cmd, queryproto.NewQueryClient, GetTickAccumulatorTrackers)
 	cmd.AddCommand(
 		osmocli.GetParams[*queryproto.ParamsRequest](
 			types.ModuleName, queryproto.NewQueryClient),
@@ -100,4 +103,31 @@ func GetTotalLiquidity() (*osmocli.QueryDescriptor, *queryproto.GetTotalLiquidit
 		Long: `{{.Short}}{{.ExampleHeader}}
 {{.CommandPrefix}} total-liquidity 1`,
 	}, &queryproto.GetTotalLiquidityRequest{}
+}
+
+func GetTickLiquidityNetInDirection() (*osmocli.QueryDescriptor, *queryproto.LiquidityNetInDirectionRequest) {
+	return &osmocli.QueryDescriptor{
+		Use:   "liquidity-net-in-direction [pool-id] [token-in-denom] [start-tick] [use-current-tick] [bound-tick] [use-no-bound]",
+		Short: "Query liquidity net in direction",
+		Long: `{{.Short}}{{.ExampleHeader}}
+{{.CommandPrefix}} 4 uosmo "[-18000000]" true "[-9000000]" true`,
+	}, &queryproto.LiquidityNetInDirectionRequest{}
+}
+
+func GetPoolAccumulatorRewards() (*osmocli.QueryDescriptor, *queryproto.PoolAccumulatorRewardsRequest) {
+	return &osmocli.QueryDescriptor{
+		Use:   "pool-accumulator-rewards [pool-id]",
+		Short: "Query pool accumulator rewards",
+		Long: `{{.Short}}{{.ExampleHeader}}
+{{.CommandPrefix}} pool-accumulator-rewards 1`,
+	}, &queryproto.PoolAccumulatorRewardsRequest{}
+}
+
+func GetTickAccumulatorTrackers() (*osmocli.QueryDescriptor, *queryproto.TickAccumulatorTrackersRequest) {
+	return &osmocli.QueryDescriptor{
+		Use:   "tick-accumulator-trackers [pool-id] [tick-index]",
+		Short: "Query tick accumulator trackers",
+		Long: `{{.Short}}{{.ExampleHeader}}
+{{.CommandPrefix}} tick-accumulator-trackers 1 "[-18000000]"`,
+	}, &queryproto.TickAccumulatorTrackersRequest{}
 }